requirements.txt
wxpython>=4.2.0
requests>=2.31.0
sqlite3
bcrypt>=4.0.1
python-dotenv>=1.0.0